requirements.txt
absl-py==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.1.0
cycler==0.11.0
dask==2022.8.1
DAWG-Python==0.7.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docopt==0.6.2
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.8.0
flatbuffers==1.12
fonttools==4.36.0
fsspec==2022.7.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
ipykernel==6.15.1
ipympl==0.9.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==3.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
nodejs==0.1.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
optional-django==0.1.0
packaging==21.3
pandarallel==1.6.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-pretrained-vit==0.0.7
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.4.0
swifter==1.3.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.12.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.1
typing_extensions==4.3.0
ua-parser==0.16.1
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==4.0.2
wrapt==1.14.1
